+++ application/src/main/java/com/kaua/ecommerce/customer/application/usecases/address/outputs/UpdateAddressOutput.java
@@ -1,0 +1,13 @@
+package com.kaua.ecommerce.customer.application.usecases.address.outputs;
+
+import com.kaua.ecommerce.customer.domain.address.Address;
+
+public record UpdateAddressOutput(
+        String id,
+        String customerId
+) {
+
+    public UpdateAddressOutput(final Address aAddress) {
+        this(aAddress.getId().value().toString(), aAddress.getCustomerId().value().toString());
+    }
+}