+++ infrastructure/src/main/java/com/kaua/ecommerce/customer/infrastructure/configurations/usecases/AddressUseCaseConfig.java
@@ -4,8 +4,10 @@ import com.kaua.ecommerce.customer.application.gateways.AddressGateway;
 import com.kaua.ecommerce.customer.application.repositories.AddressRepository;
 import com.kaua.ecommerce.customer.application.usecases.address.CreateCustomerAddressUseCase;
 import com.kaua.ecommerce.customer.application.usecases.address.UpdateAddressIsDefaultUseCase;
+import com.kaua.ecommerce.customer.application.usecases.address.UpdateAddressUseCase;
 import com.kaua.ecommerce.customer.application.usecases.address.impl.DefaultCreateCustomerAddressUseCase;
 import com.kaua.ecommerce.customer.application.usecases.address.impl.DefaultUpdateAddressIsDefaultUseCase;
+import com.kaua.ecommerce.customer.application.usecases.address.impl.DefaultUpdateAddressUseCase;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -25,5 +27,13 @@ public class AddressUseCaseConfig {
             final AddressRepository addressRepository
     ) {
         return new DefaultUpdateAddressIsDefaultUseCase(addressRepository);
+    }
+
+    @Bean
+    public UpdateAddressUseCase updateAddressUseCase(
+            final AddressRepository addressRepository,
+            final AddressGateway addressGateway
+    ) {
+        return new DefaultUpdateAddressUseCase(addressGateway, addressRepository);
     }
 }